requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
openai==1.12.0
tiktoken==0.6.0
pandas==2.2.0
python-dotenv==1.0.1
pytest==8.0.0
